rename bundle to bundle-assets
