--- cli/app/src/components/Factory/queries.ts
@@ -1,16 +1,0 @@
-import { fetchGraphQL } from "../../graphql/fetchGraphql.ts";
-
-export async function createGraph() {
-  await fetchGraphQL(`mutation CreateMany {
-    createMany(inputs: [
-      {typename:"Component"},
-      {typename:"Group"},
-      {typename:"API"},
-      {typename:"Resource"},
-      {typename:"User"},
-      {typename:"Domain"}
-    ]) {
-      id
-    }
-  }`);
-}
