+++ cli/deno.json
@@ -4,7 +4,7 @@
     "serve": "deno run -A https://deno.land/x/packup@v0.1.14/cli.ts serve app/index.html",
     "build": "deno run -A --unstable npm:vite build",
     "dev": "deno run -A ./main.ts --watch ./src -f ./example/factory.ts --app-path=./dist",
-    "bundle": "deno run -A tasks/bundle.ts",
+    "bundle-assets": "deno run -A tasks/bundle-assets.ts",
     "compile": "deno compile -A --no-check --allow-net -o graphgen ./main.ts"
   },
   "lint": {
